move frontend code to web dir
