+++ frontend/web/components/Header/Nav/MainMenuToggle/VeggieBurger.tsx
