+++ frontend/web/components/Header/Nav/SubNav/Inner.tsx
@@ -8,7 +8,7 @@ import {
     mobileMedium,
     mobileLandscape,
 } from '@guardian/pasteup/breakpoints';
-import { pillarPalette, pillarMap } from '../../../../pillars';
+import { pillarPalette, pillarMap } from '../../../../../pillars';
 
 const wrapperCollapsed = css`
     height: 36px;
