+++ frontend/web/components/Header/Nav/index.tsx
@@ -10,7 +10,7 @@ import Pillars from './Pillars';
 import MainMenuToggle from './MainMenuToggle';
 import { MainMenu } from './MainMenu';
 import SubNav from './SubNav/SubNav';
-import { getCookie } from '../../../lib/cookie';
+import { getCookie } from '../../../../lib/cookie';
 
 const centered = css`
     ${tablet} {
