+++ frontend/web/components/Picture.tsx
