+++ frontend/web/components/elements/ImageBlockComponent.tsx
