+++ frontend/web/document.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 import { extractCritical } from 'emotion-server';
 import { renderToString } from 'react-dom/server';
 
-import assets from './lib/assets';
+import assets from '../lib/assets';
 import htmlTemplate from './htmlTemplate';
 import Article from './pages/Article';
-import { GADataType } from './lib/parse-capi';
+import { GADataType } from '../lib/parse-capi';
 
 interface Props {
     data: {
